backend/requirements.txt
asgiref==3.11.0
Django==6.0
django-cors-headers==4.9.0
djangorestframework==3.16.1
pillow==12.0.0
python-decouple==3.8
sqlparse==0.5.5
